pyproject.toml
[tool.poetry]
name = "wallflower"
version = "0.1.0"
description = "A desktop wallpaper aggregator"
authors = ["alexgQQ <me@alexgrand.dev>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
praw = "^7.4.0"
numpy = "^1.21.1"
aiohttp = "^3.7.4"
aiofiles = "^0.7.0"
Pillow = "^8.3.1"
opencv-python = "^4.5.3"
imgurpython = "^1.1.7"
funcy = "^1.16"
SQLAlchemy = "^1.4.25"
PySimpleGUI = "^4.60.4"
# Pypi rel differs from main which removed numpy dep
vptree = {git = "https://github.com/RickardSjogren/vptree.git"}
# Unclear if latest is published to Pypi, grabbed freshest for some debugging
colormath = {git = "https://github.com/gtaylor/python-colormath.git"}
requests = "^2.28.2"
platformdirs = "^3.0.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
pyinstrument = "^4.0.2"
pyinstaller = "^5.8.0"
pytest = "^6.2.4"
pylint = "^2.9.6"
Faker = "^8.14.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
